@@ -12,7 +12,7 @@
     }
   ],
   "require": {
-    "php": ">=7.0"
+    "php": ">=8.0"
   },
   "require-dev": {
     "phpunit/phpunit": "^8.0",
